put functions into user.py and pet.py
